+++ CHANGES.md
@@ -1,4 +1,6 @@
-#6.0.23 ????-??-??
+#6.0.24 ????-??-??
+
+#6.0.23 2019-10-04
  - 2019-09-13 Fixed bug#[14768](https://bugs.otrs.org/show_bug.cgi?id=14768) - Advanced setting options are not reachable in System Configuration.
  - 2019-09-12 Fixed bug#[13886](https://bugs.otrs.org/show_bug.cgi?id=13886) - Access to settings with higher ConfigLevel is possible even if systemwide ConfigLevel is not matched.
  - 2019-09-11 Fixed bug#[14761](https://bugs.otrs.org/show_bug.cgi?id=14761) - Article creating with an attachment named '0' does not work and shows error message.